+++ gemfiles/Gemfile.rails-4.0.rb
@@ -18,7 +18,7 @@ group :development, :test do
   platforms :ruby, :rbx do
     gem 'sqlite3'
     gem 'mysql2', '~> 0.3.10'
-    gem 'pg'
+    gem 'pg', '~> 0.15'
     gem 'redcarpet'
   end