+++ gemfiles/Gemfile.rails-6.0.rb
@@ -2,15 +2,14 @@ source 'https://rubygems.org'
 
 gemspec path: '../'
 
-gem 'activerecord', '~> 5.0.0'
-gem 'railties', '~> 5.0.0'
-gem 'i18n', '~> 0.7.0'
+gem 'activerecord', '~> 6.0.0'
+gem 'railties', '~> 6.0.0'
 
 # Database Configuration
 group :development, :test do
   platforms :jruby do
-    gem 'activerecord-jdbcmysql-adapter', '~> 50.1'
-    gem 'activerecord-jdbcpostgresql-adapter', '~> 50.1'
+    gem 'activerecord-jdbcmysql-adapter', '~> 51.1'
+    gem 'activerecord-jdbcpostgresql-adapter', '~> 51.1'
     gem 'kramdown'
   end