+++ lib/friendly_id/slug.rb
@@ -5,6 +5,10 @@ module FriendlyId
   class Slug < ActiveRecord::Base
     belongs_to :sluggable, :polymorphic => true
 
+    def sluggable
+      sluggable_type.constantize.unscoped { super }
+    end
+
     def to_param
       slug
     end