+++ test/shared.rb
@@ -62,9 +62,9 @@ module FriendlyId
           my_model_class = Class.new(model_class)
           self.class.const_set("Foo", my_model_class)
           with_instance_of my_model_class do |record|
-            record.update_attributes my_model_class.friendly_id_config.slug_column => nil
+            record.update my_model_class.friendly_id_config.slug_column => nil
             record = my_model_class.friendly.find(record.id)
-            record.class.validate Proc.new {errors[:name] = "FAIL"}
+            record.class.validate Proc.new {errors.add(:name, "FAIL")}
             record.save
             assert_equal record.to_param, record.friendly_id
           end
@@ -132,7 +132,7 @@ module FriendlyId
         test "updating record's other values should not change the friendly_id" do
           with_instance_of model_class do |record|
             old = record.friendly_id
-            record.update_attributes! :active => false
+            record.update! active: false
             assert model_class.friendly.find old
           end
         end
@@ -174,7 +174,7 @@ module FriendlyId
         end
 
         test "should return nil for to_param with a new record" do
-          assert_equal nil, model_class.new.to_param
+          assert_nil model_class.new.to_param
         end
       end
     end