+++ test/slugged_test.rb
@@ -152,7 +152,6 @@ class SluggedTest < TestCaseClass
       assert_equal 'foo', instance.slug
     end
   end
-
 end
 
 class SlugGeneratorTest < TestCaseClass
@@ -269,6 +268,31 @@ class SlugSeparatorTest < TestCaseClass
     end
   end
 
+end
+
+class SlugLimitTest < TestCaseClass
+
+  include FriendlyId::Test
+
+  class Journalist < ActiveRecord::Base
+    extend FriendlyId
+    friendly_id :name, :use => :slugged, :slug_limit => 40
+  end
+
+  def model_class
+    Journalist
+  end
+
+  test "should limit slug size" do
+    transaction do
+      m1 = model_class.create! :name => 'a' * 50
+      assert_equal m1.slug, 'a' * 40
+      m2 = model_class.create! :name => m1.name
+      m2.save!
+      # "aaa-<uid>"
+      assert_match(/\Aa{3}\-/, m2.slug)
+    end
+  end
 end
 
 class DefaultScopeTest < TestCaseClass
@@ -389,6 +413,85 @@ class FailedValidationAfterUpdateRegressionTest < TestCaseClass
       journalist.slug_de = nil
       assert !journalist.valid?
       assert_equal "joseph-pulitzer", journalist.to_param
+    end
+  end
+
+end
+
+class ToParamTest < TestCaseClass
+
+  include FriendlyId::Test
+
+  class Journalist < ActiveRecord::Base
+    extend FriendlyId
+    validates_presence_of :active
+    friendly_id :name, :use => :slugged
+
+    attr_accessor :to_param_in_callback
+
+    after_save do
+      self.to_param_in_callback = to_param
+    end
+  end
+
+  test "to_param should return nil if record is unpersisted" do
+    assert_nil Journalist.new.to_param
+  end
+
+  test "to_param should return nil if record failed validation" do
+    journalist = Journalist.new :name => 'Clark Kent', :active => nil
+    refute journalist.save
+    assert_nil journalist.to_param
+  end
+
+  test "to_param should use slugged attribute if record saved successfully" do
+    transaction do
+      journalist = Journalist.new :name => 'Clark Kent', :active => true
+      assert journalist.save
+      assert_equal 'clark-kent', journalist.to_param
+    end
+  end
+
+  test "to_param should use original slug if existing record changes but fails to save" do
+    transaction do
+      journalist = Journalist.new :name => 'Clark Kent', :active => true
+      assert journalist.save
+      journalist.name = 'Superman'
+      journalist.slug = nil
+      journalist.active = nil
+      refute journalist.save
+      assert_equal 'clark-kent', journalist.to_param
+    end
+  end
+
+  test "to_param should use new slug if existing record changes successfully" do
+    transaction do
+      journalist = Journalist.new :name => 'Clark Kent', :active => true
+      assert journalist.save
+      journalist.name = 'Superman'
+      journalist.slug = nil
+      assert journalist.save
+      assert_equal 'superman', journalist.to_param
+    end
+  end
+
+  test "to_param should use new slug within callbacks if new record is saved successfully" do
+    transaction do
+      journalist = Journalist.new :name => 'Clark Kent', :active => true
+      assert journalist.save
+      assert_equal 'clark-kent', journalist.to_param_in_callback, "value of to_param in callback should use the new slug value"
+    end
+  end
+
+  test "to_param should use new slug within callbacks if existing record changes successfully" do
+    transaction do
+      journalist = Journalist.new :name => 'Clark Kent', :active => true
+      assert journalist.save
+      assert journalist.valid?
+      journalist.name = 'Superman'
+      journalist.slug = nil
+      assert journalist.save, "save should be successful"
+      assert_equal 'superman', journalist.to_param_in_callback, "value of to_param in callback should use the new slug value"
     end
   end